Buttholes. That's gow I feel. Merged Varun and my code, added objects to second room, tinkered a bit, I don't remember cause I'm tired.

+++ InherentGames/src/com/inherentgames/Bubble.java
@@ -76,7 +76,7 @@ public class Bubble extends WordObject{
 		objectId = id;
 	}
 	
-	public float getTimeCreated(){
+	public long getTimeCreated(){
 		return timeCreated;
 	}
 }
